requirements.txt
requests
setuptools
Pillow
python-dotenv
google-generativeai
google-genai
PyPDF2
markdown
tkhtmlview
pyrebase
firebase-admin
google-cloud-storage
urllib3==1.26.20
pycryptodome
python-docx
rich
google-generativeai